WTF - See description bellow.
I assume it has to be something else and the behaviour I'm observing is
somehow misleading me.

When the writeRecordToFile function is declared in point free notation
is does not write to the file and it doesn't fail either (at least not
that I can see). On the other hand, when adding explict parameter it
works fine.

I need to look into it further, but the type signature (infered) seemed
to be ok (observed through intero).